TEKDB/requirements.txt
#REQUIRED
confusable-homoglyphs
coverage
django >=4.2.16,<4.3
django-autocomplete-light
django-ckeditor
django-colorfield
django-leaflet
django-nested-admin
django-registration
django-reversion
django-tinymce
pillow
psycopg2-binary
psutil
django-filebrowser-no-grappelli>=4.0.0,<5.0.0
djangorestframework

XlsxWriter
#-e git+https://github.com/dominno/django-moderation.git@master#egg=moderation

#DEPENDENCIES
cryptography>=3.3.2
httplib2>=0.19.0
jinja2>=2.11.3
pyyaml>=5.4
SecretStorage>=3.3.1
testresources
twisted>=22.4.0
urllib3>=1.26.5

#OPTIONAL, BUT LOVELY FOR DEBUGGING
ipdb
ipython

# Serve static files in production containers
whitenoise>=6.0.0,<7.0.0
# FORMATTING AND LINTING
ruff
